services/integrations/app/requirements.txt
fastapi
uvicorn
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
requests
python-dotenv
slack_sdk